refactor rnns weights
